Move Utf8-Encoder from Channel to util.Utf8
