+++ src/main/java/channels/Channel.java
@@ -1,7 +1,5 @@
 package channels;
 
-import java.io.UnsupportedEncodingException;
-
 public interface Channel<T> {
     interface EventHandler {
         /**
@@ -40,19 +38,4 @@ public interface Channel<T> {
     void addEventHandler(EventHandler eventHandler);
 
     void removeEventHandler(EventHandler eventHandler);
-
-    /**
-     * UTF-8 Encoder/Decoder
-     */
-    class Encoder {
-        private final static String DEFAULT_ENCODING = "UTF-8";
-
-        public static String encodeByteArray(byte[] bytes) throws UnsupportedEncodingException {
-            return new String(bytes, DEFAULT_ENCODING);
-        }
-
-        public static byte[] decodeString(String data) throws UnsupportedEncodingException {
-            return data.getBytes(DEFAULT_ENCODING);
-        }
-    }
 }
